Added validation and open API

+++ src/main/java/com/ecommerce/product/controller/ProductController.java
@@ -32,18 +32,18 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("/ecommerce/products")
 @RequiredArgsConstructor
 public class ProductController {
-    
+
     private final ProductService service;
-    
+
     @PostMapping
     @ResponseStatus(HttpStatus.CREATED)
-    public void createProduct(@RequestBody @Valid ProductDTO productRequest){
+    public void createProduct(@RequestBody @Valid ProductDTO productRequest) {
         service.createProduct(productRequest);
     }
-    
+
     @GetMapping
     @ResponseStatus(HttpStatus.OK)
-    public List<ProductDTO> getAllProducts(){
+    public List<ProductDTO> getAllProducts() {
         return service.getAllProducts();
     }
 
@@ -55,11 +55,10 @@ public class ProductController {
         }
         return new ResponseEntity<>(productDTO, HttpStatus.OK);
     }
-    
+
     @PutMapping("/{id}")
-    @ResponseStatus(HttpStatus.OK)
     public ResponseEntity<?> update(@PathVariable(value = "id") String id,
-                    @RequestBody @Valid ProductDTO product) {
+            @RequestBody @Valid ProductDTO product) {
         ProductDTO productDTO = service.getById(id);
         if (productDTO == null) {
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
@@ -69,12 +68,16 @@ public class ProductController {
         return new ResponseEntity<>(HttpStatus.OK);
     }
 
-    @DeleteMapping("/{id}")
-    @ResponseStatus(HttpStatus.OK)
-    public void delete(@PathVariable(value = "id") String id) {
-        service.delete(id);       
-    }   
-    
+    @DeleteMapping("/{id}")    
+    public ResponseEntity<?> delete(@PathVariable(value = "id") String id) {
+        ProductDTO productDTO = service.getById(id);
+        if (productDTO == null) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+        service.delete(id);
+        return new ResponseEntity<>(HttpStatus.OK);
+    }
+
     @ResponseStatus(HttpStatus.BAD_REQUEST)
     @ExceptionHandler(MethodArgumentNotValidException.class)
     public Map<String, String> handleValidationExceptions(MethodArgumentNotValidException ex) {
